fix - corrige o bug1
O bug 1 acontece por conta da divisão de 2 inteiros, sem realizar a
conversão dessa divisão para ponto flutuante. É exigido que o resultado
seja ponto flutuante.
Também foram incluídos comentários para auxiliar no entendimento do
código, de acordo com a solicitação de projeto da Padaria.
Foi adicionado um cast double redundante nos itens vendidos por unidade,
para uma segurança extra em caso de um dos valores passar a ser por
exemplo, "4", e o bug de número inteiro ocorrer novamente.

+++ Desafio1/registradora-cheia-de-bugs/src/RelacaoPesoPreco.java
@@ -3,24 +3,29 @@ public class RelacaoPesoPreco {
     public static double retornaPrecoProduto(String item, int qtd) {
         double precoTotal = 0;
 
+        //É vendido por kilo, e cada pao pesa 60g. Cliente solicita a quantidade de paes.
         if ("pao".equals(item)) {
-            precoTotal = 12.75 * (qtd * 60 / 1000);
+            precoTotal = 12.75 * (qtd * (double) 60 / 1000);
         }
 
+        //É vendido por fatia, e cada torta tem 16 fatias. Cliente solicita a quantidade de fatias.
         if ("torta".equals(item)) {
-            precoTotal = 96.00 * (qtd / 16);
+            precoTotal = 96.00 * ( (double) qtd / 16);
         }
 
+        //Vendido por unidade. O cliente pede a quantidade.
         if ("leite".equals(item)) {
-            precoTotal = 4.48 * qtd;
+            precoTotal = (double) 4.48 * qtd;
         }
 
+        //Vendido por unidade. O cliente pede a quantidade.
         if ("cafe".equals(item)) {
-            precoTotal = 9.56 * qtd;
+            precoTotal = (double) 9.56 * qtd;
         }
 
+        //Vendido por unidade. O cliente pede a quantidade.
         if ("sanduiche".equals(item)) {
-            precoTotal = 4.5 * qtd;
+            precoTotal = (double) 4.5 * qtd;
         }
 
         return precoTotal;
